requirements.txt
numpy
pillow
scikit-learn
opencv-python
matplotlib
scipy